pyproject.toml
[build-system]
requires = ["setuptools>=60.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "fastapi-jwt"
description = "`FastAPI` extension for JTW Auth"
readme = "README.md"
license = {text = "MIT License"}
authors = [
  {name = "Konstantin Chernyshev", email = "kdchernyshev@gmail.com"},
]
dynamic = ["version"]

requires-python = ">=3.8"
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Typing :: Typed"
]

dependencies = [
    "fastapi >=0.50.0",
]


[project.urls]
homepage = "https://github.com/k4black/fastapi-jwt"
documentation = "https://k4black.github.io/fastapi-jwt/"


[project.optional-dependencies]
authlib = [
    "Authlib >=1.3.0"
]
python_jose = [
    "python-jose[cryptography] >=3.3.0"
]
test = [
    "Authlib >=1.3.0",
    "python-jose[cryptography] >=3.3.0",
    "httpx >=0.23.0,<1.0.0",
    "pytest >=7.0.0,<9.0.0",
    "pytest-cov >=4.0.0,<5.0.0",
    "pytest-mock >=3.0.0,<4.0.0",
    "requests >=2.28.0,<3.0.0",
    "black ==24.2.0",
    "mypy >=1.0.0,<2.0.0",
    "flake8 >=6.0.0,<8.0.0",
    "ruff >=0.1.0,<1.0.0",
    "isort >=5.11.0,<6.0.0",
    "types-python-jose ==3.3.4.8"
]
docs = [
    "mkdocs >=1.4.0,<2.0.0",
    "mkdocs-material >=9.0.0,<10.0.0",
    "MkAutoDoc >=0.2.0,<1.0.0",
    "lazydocs >=0.4.5,<1.0.0",
    "mkdocs-include-markdown-plugin >=4.0.0,<7.0.0",
    "mkdocs-awesome-pages-plugin >=2.8.0,<3.0.0",
    "mike >=1.1.0,<3.0.0"
]


[tool.setuptools.dynamic]
version = {file = "VERSION"}


[mypy]
ignore_missing_imports = true
no_incremental = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = false
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
allow_untyped_decorators = true


[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=fastapi_jwt/ --cov-report term-missing"


# TODO: remove in favor of ruff "I" ruleset
[tool.isort]
profile = "black"
src_paths = ["fastapi-jwt", "tests"]
known_first_party = ["fastapi_jwt", "tests"]
line_length = 120
combine_as_imports = true


[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []

fixable = ["ALL"]  # Allow fix for all enabled rules (when `--fix`) is provided.
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
